# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Documentation for the Application API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Server #1"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "register"
                ]
            }
        },
        "/login": {
            "post": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "login"
                ]
            }
        },
        "/logout": {
            "post": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "logout"
                ]
            }
        },
        "/user": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        }
    },
    "tags": [],
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "password": {
                        "type": "string"
                    },
                    "remember_token": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "email",
                    "password"
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}